Split the code for countdown timer
