+++ main/gui/lvgl_manager.cpp
@@ -11,6 +11,7 @@
 
 // project header files
 #include "battery_status.hpp"
+#include "countdown_timer.hpp"
 #include "lvgl_manager.hpp"
 #include "sdkconfig.h"
 
@@ -24,34 +25,10 @@ SemaphoreHandle_t xGuiSemaphore;
 // constants
 static const uint32_t kLvTickPeriodMs = 1;
 
-static void set_angle(void *obj, int32_t v) {
-  lv_arc_set_value(static_cast<lv_obj_t *>(obj), v);
-}
-
-void drawCountDownTimer() {
-  // draw progress arc
-  lv_obj_t *progress_arc = lv_arc_create(lv_scr_act());
-  lv_arc_set_rotation(progress_arc, 270);
-  lv_arc_set_bg_angles(progress_arc, 0, 360);
-  lv_obj_remove_style(progress_arc, NULL, LV_PART_KNOB);
-  lv_obj_center(progress_arc);
-
-  lv_anim_t anim;
-  lv_anim_init(&anim);
-  lv_anim_set_var(&anim, progress_arc);
-  lv_anim_set_exec_cb(&anim, set_angle);
-  lv_anim_set_time(&anim, 1000);
-  lv_anim_set_repeat_count(&anim, LV_ANIM_REPEAT_INFINITE);    /*Just for the demo*/
-  lv_anim_set_repeat_delay(&anim, 500);
-  lv_anim_set_values(&anim, 0, 100);
-  lv_anim_start(&anim);
-}
-
 static void lv_tick_task(void *args) {
   (void) args;
   lv_tick_inc(kLvTickPeriodMs);
 }
-
 
 void guiTask(void *pvParameter) {
   (void) pvParameter;
@@ -68,7 +45,7 @@ void guiTask(void *pvParameter) {
 
   // gui code
   std::thread battery_status_thread(drawBatteryPercentage);
-  drawCountDownTimer();
+  drawCountdownTimer();
   
   // forever loop
   while (1) {
